Split javascript to ease reuse. Remove hardcoding.
On bbcmicro.co.uk I wanted a branded version of the index page which
doesn't display the catalogue of discs. To allow this type of reuse
and customisation I:

 - Split out the javascript into a seperate file

 - Removed the hard coded directory name 'disks' from the javascript

 - Added the subdirectory to the disks.json

 - Removed the directory from the display on the details page

The page works identically, apart from the disc details page
which does display the directory. I thought this might be useful,
so left this as it was though it would be possible to remove this
if necessary.